Remove the first grid when creating the custom one
So, I finally found a solution for that, I created a variable called "verify" inside the function gridCreator, that references the .grid class (each block has a .grid class), and verifies if the number of the blocks are superior than the value * value + value (rows * columns + each row) inputted by the user. If that condition is true, the container variable is called, and while the firstChild (.grid class is child of container) of this variable is true, the firstChild will be removed, until firstChild is false, and finally executes the function gridCreator again.